+++ smithy-model/src/main/java/software/amazon/smithy/model/validation/validators/TraitConflictValidator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -16,6 +16,7 @@
 package software.amazon.smithy.model.validation.validators;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -25,6 +26,7 @@ import software.amazon.smithy.model.Model;
 import software.amazon.smithy.model.shapes.Shape;
 import software.amazon.smithy.model.shapes.ShapeId;
 import software.amazon.smithy.model.traits.Trait;
+import software.amazon.smithy.model.traits.TraitDefinition;
 import software.amazon.smithy.model.validation.AbstractValidator;
 import software.amazon.smithy.model.validation.ValidationEvent;
 
@@ -35,29 +37,48 @@ public final class TraitConflictValidator extends AbstractValidator {
 
     @Override
     public List<ValidationEvent> validate(Model model) {
-        return model.shapes()
-                .flatMap(shape -> {
-                    // Map of trait shape IDs to trait value.
-                    Map<ShapeId, Trait> traits = shape.getAllTraits();
-                    // Map of trait shape ID to a list of found conflicting traits.
-                    Map<ShapeId, List<ShapeId>> conflicts = new HashMap<>();
+        // Find all trait definitions and collect conflicting traits into a map.
+        Map<Shape, Map<ShapeId, List<ShapeId>>> shapeToTraitConflicts = new HashMap<>();
+        for (Shape shape : model.getShapesWithTrait(TraitDefinition.class)) {
+            TraitDefinition trait = shape.expectTrait(TraitDefinition.class);
+            // Only look at trait definitions that define conflicting traits.
+            if (!trait.getConflicts().isEmpty()) {
+                findAndCollectConflicts(model, shape.getId(), trait.getConflicts(), shapeToTraitConflicts);
+            }
+        }
 
-                    traits.forEach((k, v) -> {
-                        model.getTraitDefinition(v.toShapeId()).ifPresent(definition -> {
-                            definition.getConflicts().forEach(conflict -> {
-                                if (traits.containsKey(conflict)) {
-                                    conflicts.computeIfAbsent(k, key -> new ArrayList<>()).add(conflict);
-                                }
-                            });
-                        });
-                    });
+        if (shapeToTraitConflicts.isEmpty()) {
+            return Collections.emptyList();
+        }
 
-                    return conflicts.isEmpty() ? Stream.empty() : Stream.of(foundConflicts(shape, conflicts));
-                })
-                .collect(Collectors.toList());
+        List<ValidationEvent> events = new ArrayList<>();
+        for (Map.Entry<Shape, Map<ShapeId, List<ShapeId>>> entry : shapeToTraitConflicts.entrySet()) {
+            events.add(emitForConflicts(entry.getKey(), entry.getValue()));
+        }
+
+        return events;
     }
 
-    private ValidationEvent foundConflicts(Shape shape, Map<ShapeId, List<ShapeId>> conflicts) {
+    // Find shapes that use this trait and also apply conflicting traits.
+    private void findAndCollectConflicts(
+            Model model,
+            ShapeId trait,
+            List<ShapeId> conflicts,
+            Map<Shape, Map<ShapeId, List<ShapeId>>> shapeToTraitConflicts
+    ) {
+        for (Shape shape : model.getShapesWithTrait(trait)) {
+            for (ShapeId conflict : conflicts) {
+                if (shape.hasTrait(conflict)) {
+                    shapeToTraitConflicts
+                            .computeIfAbsent(shape, id -> new HashMap<>())
+                            .computeIfAbsent(trait, id -> new ArrayList<>())
+                            .add(conflict);
+                }
+            }
+        }
+    }
+
+    private ValidationEvent emitForConflicts(Shape shape, Map<ShapeId, List<ShapeId>> conflicts) {
         return error(shape, "Found conflicting traits on " + shape.getType() + " shape: "
                             + conflicts.entrySet().stream()
                                     .flatMap(this::lines)